@robodoo 2025-05-07
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/208875.png)](https://mergebot.odoo.com/odoo/odoo/pull/208875)